Add privacy dropdown client component
