+++ src/app/(routes)/dashboard/photos/components/PhotoRow/PhotoRow.tsx
@@ -1,8 +1,3 @@
-"use client"
-
-// Hooks
-import { useState } from "react"
-
 // Types
 import { Photo } from "@/types/models"
 
@@ -10,35 +5,18 @@ import { Photo } from "@/types/models"
 import Link from "next/link"
 import Thumbnail from "./Thumbnail"
 import DeleteRowButton from "./DeleteRowButton"
-
-// Services
-import { updatePhoto } from "@/services/photoService"
+import PrivacyDropdown from "./PrivacyDropdown"
 
 // Helpers
 import { formatBytes } from "@/lib/helpers"
 
 const PhotoRow = ({ photo }: { photo: Photo }) => {
-
   const path = `/dashboard/photos/${photo.id}`
-
-  const [privacySetting, setPrivacySetting] = useState(
-    photo.isPublic ? "Public" : "Private"
-  )
 
   // Need to place the following UI:
   const viewBtn = <Link href={path}>VIEW</Link>
   const checkBox = <input type="checkbox" readOnly checked={photo.isUploaded} />
 
-  // TODO Add modal to view image src, details and stats
-
-  const handlePrivacyChange = async (e: React.ChangeEvent<HTMLSelectElement>) => {
-    console.log("run")
-    setPrivacySetting(e.target.value)
-    const formData = { isPublic: e.target.value === "Public" }
-    const res = await updatePhoto(photo.id, formData)
-
-    console.log(res)
-  }
 
   return (
     <div className="row-container">
@@ -48,16 +26,7 @@ const PhotoRow = ({ photo }: { photo: Photo }) => {
         <p>{photo.fileSize && formatBytes(photo.fileSize)}</p>
       </div>
 
-      <select
-        id="isPublic"
-        name="isPublic"
-        value={privacySetting}
-        onChange={handlePrivacyChange}
-        className="row-ui-element pl-1"
-      >
-        <option value="Public">Public</option>
-        <option value="Private">Private</option>
-      </select>
+      <PrivacyDropdown photo={photo} />
 
       <DeleteRowButton photoId={photo.id} />
     </div>
